Update access (#35)
* Added conditions to views

Improved information on views
Added username to login/logout status
Added nav-bar to login-page
Changed access-rights for endpoints

* Updated access permissions

* Fixed permission issue

* Fixed exception throws for endpoints and updated download permissions

* Added access checks to endpoints

* Removed unnecessary todo

+++ src/main/java/io/fi0x/languagegenerator/rest/LanguageController.java
@@ -9,12 +9,14 @@ import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.SessionAttributes;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.InvalidObjectException;
 import java.util.Objects;
@@ -22,7 +24,7 @@ import java.util.Objects;
 @Slf4j
 @Controller
 @AllArgsConstructor
-@SessionAttributes({"language", "amount", "words"})
+@SessionAttributes({"language", "amount", "words", "username"})
 public class LanguageController
 {
     private GenerationService generationService;
@@ -46,7 +48,15 @@ public class LanguageController
         else
             model.put("amount", amount);
 
-        model.put("languageName", languageService.getLanguageData(language).getName());
+        LanguageData languageData = languageService.getLanguageData(language);
+
+        if(!languageData.getUsername().equals(authenticationService.getAuthenticatedUsername()) && !languageData.isVisible())
+        {
+            log.info("User '{}' tried to generate words with language {}, to which he has no access to", authenticationService.getAuthenticatedUsername(), language);
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to access the selected language");
+        }
+
+        model.put("languageName", languageData.getName());
 
         try
         {
@@ -91,7 +101,15 @@ public class LanguageController
     {
         log.info("editLanguage() called");
 
-        model.put("languageData", languageService.getLanguageData(languageId));
+        LanguageData languageData = languageService.getLanguageData(languageId);
+
+        if(!languageData.getUsername().equals(authenticationService.getAuthenticatedUsername()))
+        {
+            log.info("User '{}' tried to edit language {}, to which he has no access to", authenticationService.getAuthenticatedUsername(), languageData.getId());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to edit the selected language");
+        }
+
+        model.put("languageData", languageData);
 
         return "language";
     }
@@ -103,13 +121,19 @@ public class LanguageController
         log.info("addLanguage() called");
         log.debug("arguments for addLanguage() call are model={}, languageData={}", model, languageData);
 
+        if(!languageData.getUsername().equals(authenticationService.getAuthenticatedUsername()))
+        {
+            log.info("User '{}' tried to update language {}, to which he has no access to", authenticationService.getAuthenticatedUsername(), languageData.getId());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to change the selected language");
+        }
+
         try
         {
             languageService.addLanguage(languageData);
         } catch (InvalidObjectException e)
         {
             log.info("Could not save the language because it was not complete.");
-            return "redirect:language";
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not save the language, because the language-object was not complete", e);
         }
 
         return "redirect:/";
@@ -121,10 +145,19 @@ public class LanguageController
     {
         log.info("deleteLanguage() called for languageId={}", languageId);
 
+        if(!authenticationService.getAuthenticatedUsername().equals(languageService.getLanguageCreator(languageId)))
+        {
+            log.info("User '{}' tried to delete language {}, to which he has no access to", authenticationService.getAuthenticatedUsername(), languageId);
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You are not allowed to delete the selected language");
+        }
+
         if(languageService.deleteLanguage(languageId))
             log.info("Language with id={} successfully deleted", languageId);
         else
+        {
             log.warn("Could not delete language with id={}", languageId);
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "The language you were trying to delete, could not be found");
+        }
 
         return "redirect:/";
     }
